day 9: added middleware and authorization for both sign in and sign up, redirects if user is signed in or not based on route

+++ routes.php
@@ -13,18 +13,20 @@
 $router->get("/", "HomeController@index");
 
 $router->get("/listings", "ListingController@index");
-$router->get("/listings/create", "ListingController@create");
+$router->get("/listings/create", "ListingController@create", ["auth"]);
 $router->get("/listings/{id}", "ListingController@show");
 $router->put("/listings/{id}", "ListingController@update");
-$router->get("/listings/edit/{id}", "ListingController@edit");
+$router->get("/listings/edit/{id}", "ListingController@edit", ["auth"]);
 
-$router->post("/listings", "ListingController@store");
-$router->delete("/listings/{id}", "ListingController@destroy");
+$router->post("/listings", "ListingController@store", ["auth"]);
+$router->delete("/listings/{id}", "ListingController@destroy", ["auth"]);
 
-$router->get("/auth/register", "UserController@create");
-$router->get("/auth/login", "UserController@login");
+$router->get("/auth/register", "UserController@create", ["guest"]);
+$router->get("/auth/login", "UserController@login", ["guest"]);
 
-$router->post("/auth/register", "UserController@store");
+$router->post("/auth/register", "UserController@store", ["guest"]);
+$router->post("/auth/login", "UserController@authenticate", ["guest"]);
+$router->post("/auth/logout", "UserController@logout", ["auth"]);
 
 
 // $router->get("/listings/create", "controllers/listings/create.php");
